+++ ops/testnet/staging/backend/config.tf
@@ -23,6 +23,7 @@ locals {
     chains = {
       "1111" = {}
       "2221" = {}
+      "3331" = {}
     }
     environment = var.stage
   })