+++ ops/testnet/staging/core/config.tf
@@ -52,6 +52,15 @@ locals {
         ]
         providers = ["https://eth-kovan.alchemyapi.io/v2/${var.kovan_alchemy_key_0}"]
       }
+      "3331" = {
+        providers = ["https://eth-goerli.alchemyapi.io/v2/${var.kovan_alchemy_key_0}", "https://rpc.ankr.com/eth_goerli"]
+        assets = [
+          {
+            name    = "TEST"
+            address = "0xD426e23A6a9524101CDC017e01dDc3262B7aA65D"
+          }
+        ]
+      }
     }
 
     environment = var.stage
@@ -90,6 +99,15 @@ locals {
           }
         ]
       }
+      "3331" = {
+        providers = ["https://eth-goerli.alchemyapi.io/v2/${var.kovan_alchemy_key_1}", "https://rpc.ankr.com/eth_goerli"]
+        assets = [
+          {
+            name    = "TEST"
+            address = "0xD426e23A6a9524101CDC017e01dDc3262B7aA65D"
+          }
+        ]
+      }
     }
     web3SignerUrl = "https://${module.web3signer.service_endpoint}"
     environment   = var.stage
@@ -107,6 +125,9 @@ locals {
       }
       "2221" = {
         providers = ["https://eth-kovan.alchemyapi.io/v2/${var.kovan_alchemy_key_1}"]
+      }
+      "3331" = {
+        providers = ["https://eth-goerli.alchemyapi.io/v2/${var.kovan_alchemy_key_1}", "https://rpc.ankr.com/eth_goerli"]
       }
     }
     environment = var.stage