+++ packages/adapters/subgraph/src/lib/helpers/parse.ts
@@ -45,6 +45,8 @@ export const originTransfer = (entity: any): OriginTransfer => {
     xparams: {
       to: entity.to,
       callData: entity.callData,
+      forceSlow: entity.forceSlow,
+      receiveLocal: entity.receiveLocal,
     },
 
     // Origin Info
@@ -122,6 +124,8 @@ export const destinationTransfer = (entity: any): DestinationTransfer => {
         ? {
             to: entity.to,
             callData: entity.callData,
+            forceSlow: entity.forceSlow,
+            receiveLocal: entity.receiveLocal,
           }
         : undefined,