+++ packages/deployments/contracts/contracts_forge/ConnextHandler.t.sol
@@ -284,7 +284,9 @@ contract ConnextHandlerTest is ForgeHelper {
       to,
       bytes("0x"),
       domain,
-      destinationDomain
+      destinationDomain,
+      false,
+      false
     );
     IConnextHandler.XCallArgs memory args = IConnextHandler.XCallArgs(
       callParams,
@@ -311,7 +313,14 @@ contract ConnextHandlerTest is ForgeHelper {
     uint256 relayerFee = 0.01 ether;
     address transactingAssetId = address(originAdopted);
 
-    IConnextHandler.CallParams memory callParams = IConnextHandler.CallParams(to, bytes(""), domain, destinationDomain);
+    IConnextHandler.CallParams memory callParams = IConnextHandler.CallParams(
+      to,
+      bytes(""),
+      domain,
+      destinationDomain,
+      false,
+      false
+    );
     IConnextHandler.XCallArgs memory args = IConnextHandler.XCallArgs(
       callParams,
       transactingAssetId,
@@ -326,7 +335,7 @@ contract ConnextHandlerTest is ForgeHelper {
     // TODO Correctly calculate the message
     // Harcoded the message from the emitted event since here we are only testing that relayerFee is included
     bytes
-      memory message = hex"00000001000000000000000000000000c94cf1a6d4b8a25e424b3ed8792eed1f1b95b86e030000000000000000000000000000000000000000000000000000000000000064000000000000000000000000000000000000000000000000000000000000000020b4b2eeb4ea213a5e7d1e1d2a3a1a437fbe7c8b3490898b0474b0fe66dda70aca0184c1e32ae98daca86416c9ece9d32771270d0b1ef32fb55bf30918e8cc7b";
+      memory message = hex"00000001000000000000000000000000c94cf1a6d4b8a25e424b3ed8792eed1f1b95b86e030000000000000000000000000000000000000000000000000000000000000064000000000000000000000000000000000000000000000000000000000000000020b4b2eeb4ea213a5e7d1e1d2a3a1a437fbe7c8b3490898b0474b0fe66dda70a747136416651ffe40851fbe308e045079987fc3f10ccfe4b215867a350d94d4e";
 
     // NOTE: the `amount` and `bridgedAmt` are 0 because `.balanceOf` of the origin asset returns
     // 0 always via setup function
@@ -353,7 +362,9 @@ contract ConnextHandlerTest is ForgeHelper {
       to,
       bytes("0x"),
       domain,
-      destinationDomain
+      destinationDomain,
+      false,
+      false
     );
     IConnextHandler.XCallArgs memory args = IConnextHandler.XCallArgs(
       callParams,
@@ -384,7 +395,9 @@ contract ConnextHandlerTest is ForgeHelper {
       to,
       bytes("0x"),
       domain,
-      destinationDomain
+      destinationDomain,
+      false,
+      false
     );
     IConnextHandler.XCallArgs memory args = IConnextHandler.XCallArgs(
       callParams,
@@ -413,7 +426,9 @@ contract ConnextHandlerTest is ForgeHelper {
       to,
       bytes("0x"),
       domain,
-      destinationDomain
+      destinationDomain,
+      false,
+      false
     );
     IConnextHandler.XCallArgs memory args = IConnextHandler.XCallArgs(
       callParams,
@@ -448,7 +463,9 @@ contract ConnextHandlerTest is ForgeHelper {
       to,
       bytes("0x"),
       domain,
-      destinationDomain
+      destinationDomain,
+      false,
+      false
     );
     IConnextHandler.XCallArgs memory args = IConnextHandler.XCallArgs(
       callParams,
@@ -478,7 +495,9 @@ contract ConnextHandlerTest is ForgeHelper {
       to,
       bytes("0x"),
       domain,
-      destinationDomain
+      destinationDomain,
+      false,
+      false
     );
     IConnextHandler.XCallArgs memory args = IConnextHandler.XCallArgs(
       callParams,
@@ -512,7 +531,14 @@ contract ConnextHandlerTest is ForgeHelper {
     uint256 relayerFee = 0.01 ether;
     address transactingAssetId = address(originAdopted);
 
-    IConnextHandler.CallParams memory callParams = IConnextHandler.CallParams(to, bytes(""), domain, destinationDomain);
+    IConnextHandler.CallParams memory callParams = IConnextHandler.CallParams(
+      to,
+      bytes(""),
+      domain,
+      destinationDomain,
+      false,
+      false
+    );
     IConnextHandler.XCallArgs memory args = IConnextHandler.XCallArgs(
       callParams,
       transactingAssetId,